CHG | REFAC | Update PropTypes for Toggler component and rename custom hook path
