--- src/components/imagePlaceholder/canvasHolder/canvasHolder.css
@@ -1,7 +1,0 @@
-.canvas-back {
-    /*position: absolute;*/
-}
-
-.canvas-fore {
-    position: absolute;
-}
