+++ src/hooks/calcCanvasDim.js
